+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 7.3.0 (December 22, 2025)
+
+FEATURES:
+
+* Add `selectel_private_dns_service_v1` resource ([#367](https://github.com/selectel/terraform-provider-selectel/pull/367))
+* Add documentation for datasources and resource above ([#367](https://github.com/selectel/terraform-provider-selectel/pull/367))
+* Add Cloud Private DNS API SDK — [private-dns-go](https://github.com/selectel/private-dns-go) ([#367](https://github.com/selectel/terraform-provider-selectel/pull/367))
+
 ## 7.2.0 (November 18, 2025)
 
 FEATURES: